Configurable models as well as fix render bugs for static assets
