+++ pyramid_precis_creative_template/models.py
@@ -1,4 +1,6 @@
-from typing import Optional
+from typing import Dict, List, Optional, Union
+
+from pydantic import Field
 from pydantic_yaml import YamlModel
 
 
@@ -31,3 +33,26 @@ class ContactInfo(YamlModel):
     @property
     def linkedin_short(self):
         return self.linkedin.replace("https://", "").replace("http://", "")
+
+
+class Position(YamlModel):
+    title: str
+    employer: str
+    date: DateRange
+    description: str
+    bullets: List[Union[str, Dict[str, str]]] = Field(default_factory=list)
+
+
+class Education(YamlModel):
+    degree_date: str
+    degree_name: str
+    degree_concentration: str
+    institution_name: str
+
+
+class Resume(YamlModel):
+    contact_info: ContactInfo
+    executive_summary: str
+    qualifications: Union[Dict[str, List[str]], List[Dict]]
+    positions: List[Position]
+    educations: List[Education]
