+++ pyramid_precis_creative_template/templates/resume.jinja2
@@ -8,7 +8,7 @@
         <title>{{ user_name }} Resume</title>
     </head>
     
-    <link href="{{request.static_url('resume:static/favicon_io/favicon.ico')}}" rel="shortcut icon">
+    <link href="{{request.static_url('pyramid_precis_creative_template:static/favicon_io/favicon.ico')}}" rel="shortcut icon">
     
     <!-- Bootstrap core CSS -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
@@ -16,21 +16,21 @@
     <!-- Custom Fonts -->
     <link href="https://fonts.googleapis.com/css?family=Monoton|Anton|Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800"
           rel="stylesheet">
-    <link href="{{request.static_url('resume:static/creative/vendor/icomoon/style.css')}}" rel="stylesheet" type="text/css">
+    <link href="{{request.static_url('pyramid_precis_creative_template:static/vendor/icomoon/style.css')}}" rel="stylesheet" type="text/css">
     <!-- Custom Fonts -->
-    <link href="{{request.static_url('resume:static/creative/vendor/font-awesome/css/font-awesome.min.css')}}" rel="stylesheet" type="text/css">
+    <link href="{{request.static_url('pyramid_precis_creative_template:static/vendor/font-awesome/css/font-awesome.min.css')}}" rel="stylesheet" type="text/css">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/solid.css"
           integrity="sha384-wnAC7ln+XN0UKdcPvJvtqIH3jOjs9pnKnq9qX68ImXvOGz2JuFoEiCjT8jyZQX2z" crossorigin="anonymous">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/brands.css"
           integrity="sha384-nT8r1Kzllf71iZl81CdFzObMsaLOhqBU1JD2+XoAALbdtWaXDOlWOZTR4v1ktjPE" crossorigin="anonymous">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/fontawesome.css"
           integrity="sha384-HbmWTHay9psM8qyzEKPc8odH4DsOuzdejtnr+OFtDmOcIVnhgReQ4GZBH7uwcjf6" crossorigin="anonymous">
-    <link href="{{request.static_url('resume:static/creative/vendor/icomoon/style.css')}}" rel="stylesheet" type="text/css">
+    <link href="{{request.static_url('pyramid_precis_creative_template:static/vendor/icomoon/style.css')}}" rel="stylesheet" type="text/css">
     
     <!-- Theme CSS -->
-    <link href="{{request.static_url('resume:static/creative/css/creative.css')}}?v=1.0.2" rel="stylesheet">
-    <link href="{{request.static_url('resume:static/creative/css/theme.css')}}?v=1.0.4" rel="stylesheet">
-    <link href="{{request.static_url('resume:static/creative/css/resume.css')}}" rel="stylesheet">
+    <link href="{{request.static_url('pyramid_precis_creative_template:static/css/creative.css')}}?v=1.0.2" rel="stylesheet">
+    <link href="{{request.static_url('pyramid_precis_creative_template:static/css/theme.css')}}?v=1.0.4" rel="stylesheet">
+    <link href="{{request.static_url('pyramid_precis_creative_template:static/css/resume.css')}}" rel="stylesheet">
     
     <!-- Custom fonts for this template -->
     <link href="https://fonts.googleapis.com/css?family=Saira+Extra+Condensed:100,200,300,400,500,600,700,800,900"
@@ -38,7 +38,7 @@
     <link href="https://fonts.googleapis.com/css?family=Open+Sans:300,300i,400,400i,600,600i,700,700i,800,800i"
           rel="stylesheet">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/devicons/devicon@v2.15.1/devicon.min.css">
-    <link href="{{request.static_url('resume:static/creative/vendor/devicon/devicon.css') }}" rel="stylesheet">
+    <link href="{{request.static_url('pyramid_precis_creative_template:static/vendor/devicon/devicon.css') }}" rel="stylesheet">
 
     <div id="page">
         <div id="resume">
